+++ tests/fixtures/visualization/stage_3.py
@@ -1,0 +1,6 @@
+def configure(context):
+    context.stage("tests.fixtures.visualization.stage_1")
+    context.stage("tests.fixtures.visualization.stage_2")
+
+def execute(context):
+    return